Support for DMA data transfers and improved routing
Support the new DMA nodes, which transfer data at DDR speeds and
avoid using the CPU for data transfers (at the cost of latency).
This is functional with the latest logic, but still in development.

Routing can now also be done through file handles, one can instruct
a fifo (cpu or dma type) to create a route from any other node to or
from this fifo, and each node or fifo can be queried for its ID.